Former-commit-id: 19279049beb052e3fd720277f7b7348eac60b3df

+++ sql/tool/docker/cajamar/execute-scripts.sh
@@ -3,6 +3,9 @@
 CONTAINER_NAME=$1
 DOCKER_INNER_ERROR_LOG=$2
 
+
+PACKAGE_DIR=/setup/package
+
 if [[ "x$CONTAINER_NAME" == "x" || "x$DOCKER_INNER_ERROR_LOG" == "x" ]]; then
 	echo "ERROR: No se puede continuar con la ejecución de scripts de la BBDD"
 	echo "ERROR: Uso: $0 CONTAINER_NAME DOCKER_INNER_ERROR_LOG"
@@ -29,26 +32,31 @@ function log_script_output () {
 	done
 }
 
-echo "<Docker [$CONTAINER_NAME]>: Ejecutando scripts..."
-export PATH=$PATH:$ORACLE_HOME/bin
-export NLS_LANG=$CUSTOM_NLS_LANG
-echo "export NLS_LANG=$NLS_LANG" >> /home/oracle/.bashrc
-echo "<Docker [$CONTAINER_NAME]>: NLS_LANG=$NLS_LANG"
-cd /sql-package/DDL
-rm -Rf *.log
-./DDL-scripts.sh admin@orcl admin@orcl
-err_code=$?
-log_script_output
-if [[ $err_code -ne 0 ]]; then
-	echo "<Docker [$CONTAINER_NAME]>: Abortando por errores"
-	exit 1
-fi
-cd /sql-package/DML
-rm -Rf *.log
-./DML-scripts.sh admin@orcl admin@orcl
-err_code=$?
-log_script_output
-if [[ $err_code -ne 0 ]]; then
-	echo "<Docker [$CONTAINER_NAME]>: Abortando por errores"
+if [[ -d $PACKAGE_DIR ]]; then
+	echo "<Docker [$CONTAINER_NAME]>: Ejecutando scripts..."
+	export PATH=$PATH:$ORACLE_HOME/bin
+	export NLS_LANG=$CUSTOM_NLS_LANG
+	echo "export NLS_LANG=$NLS_LANG" >> /home/oracle/.bashrc
+	echo "<Docker [$CONTAINER_NAME]>: NLS_LANG=$NLS_LANG"
+	cd $PACKAGE_DIR/DDL
+	rm -Rf *.log
+	./DDL-scripts.sh admin@orcl admin@orcl
+	err_code=$?
+	log_script_output
+	if [[ $err_code -ne 0 ]]; then
+		echo "<Docker [$CONTAINER_NAME]>: Abortando por errores"
+		exit 1
+	fi
+	cd $PACKAGE_DIR/DML
+	rm -Rf *.log
+	./DML-scripts.sh admin@orcl admin@orcl
+	err_code=$?
+	log_script_output
+	if [[ $err_code -ne 0 ]]; then
+		echo "<Docker [$CONTAINER_NAME]>: Abortando por errores"
+		exit 1
+	fi
+else
+	echo "<Docker [$CONTAINER_NAME] ERROR> $PACKAGE_DIR no existe."
 	exit 1
 fi
